Reorganizing the include files and namespace
