+++ include/af/nn.h
@@ -9,8 +9,8 @@
 
 #pragma once
 
-#include <afml/util/common.hpp>
-#include <afml/nn/Nodes.hpp>
-#include <afml/nn/Weights.hpp>
-#include <afml/nn/Activations.hpp>
-#include <afml/nn/Networks.hpp>
+#include <af/nn/common.hpp>
+#include <af/nn/Nodes.hpp>
+#include <af/nn/Weights.hpp>
+#include <af/nn/Activations.hpp>
+#include <af/nn/Networks.hpp>
