+++ include/af/nn/Activations/Activation.hpp
@@ -8,10 +8,10 @@
  ********************************************************/
 #pragma once
 
-#include <afml/util/common.hpp>
-#include <afml/nn/Nodes/Node.hpp>
+#include <af/nn/common.hpp>
+#include <af/nn/Nodes/Node.hpp>
 
-namespace afml
+namespace af
 {
 
     namespace nn
