+++ include/af/nn/Nodes/Linear.hpp
@@ -8,11 +8,11 @@
  ********************************************************/
 #pragma once
 
-#include <afml/util/common.hpp>
-#include <afml/nn/Weights.hpp>
-#include <afml/nn/Nodes/Node.hpp>
+#include <af/nn/common.hpp>
+#include <af/nn/Weights.hpp>
+#include <af/nn/Nodes/Node.hpp>
 
-namespace afml
+namespace af
 {
     namespace nn
     {
