--- include/afml/nn/Nodes.hpp
@@ -1,12 +1,0 @@
-/*******************************************************
- * Copyright (c) 2015, ArrayFire
- * All rights reserved.
- *
- * This file is distributed under 3-clause BSD license.
- * The complete license agreement can be obtained at:
- * http://arrayfire.com/licenses/BSD-3-Clause
- ********************************************************/
-#pragma once
-
-#include <afml/nn/Nodes/Node.hpp>
-#include <afml/nn/Nodes/Linear.hpp>
